Initial commit
"My First Bootstrap Project: Website Replica"
This is a small website I created as part of learning Bootstrap for the first time, a popular front-end framework, to build a responsive website. It is a replica of a website from a tech YouTube tutorial, where I followed along to understand the basics of layout, responsiveness, and design using Bootstrap. It's a simple 3-4 page website, showcasing the core concepts of web development with Bootstrap.
Learning how to use Bootstrap’s grid system and components to create a responsive and functional website. Practicing responsive web design, ensuring that the site works across different devices. Understanding how to incorporate pre-built components from Bootstrap and customize them.
This project helped me in understanding of front-end web development, and it’s an important step in my journey to becoming a proficient web developer. Still there's more to explore and accumulate.
This website was built by following a tutorial by CodeWithHarry. The design is a replica for learning purposes. All credit for the original design and tutorial goes to him.